Add influence of families (home) on incertitude and belief
For now the parameters are not bad, the influence decrease and the
entropy as well. At a certain point, the entropy begins to increase,
maybe due to the high incertitude. The beliefs tend to approach 1 over
time on all the simulation i made. Also the less represented belief
intervals are "swallowed" by the bigger. Finally maybe the agents' color
is not enough variated, because we see a lot of yellow.